pyproject.toml
[project]
name = "pcjam"
version = "0.1.0"
description = "Predictive coding grid that learns to generate instrument parts from MIDI"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15,<4.0",
    "django-cors-headers>=4.3,<5.0",
    "pretty-midi>=0.2.10",
    "mido>=1.3",
    "numpy>=1.26",
]

[project.optional-dependencies]
cuda = ["jax[cuda12]>=0.4.30"]
cpu = ["jax>=0.4.30"]
dev = ["pytest>=8.0", "pytest-django>=4.8"]
